LG-12039: pull in changes from main
